pyproject.toml
[project]
name = "order-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "prisma>=0.15.0",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "supabase>=2.27.3",
]
